+++ PrayerApp/Services/ISettings.cs
@@ -15,5 +15,4 @@ public interface ISettings
     bool AllowNotifications { get; set; }
     bool QuickAddTipDismissed { get; set; }
     bool PrayerTimeLandscape { get; set; }
-    bool HideSharedWithMe { get; set; }
 }